requirements.txt
python-dotenv
ytmusicapi